backend/requirements/common.txt
Django==3.1.7
pillow
djangorestframework
djangorestframework-jwt
django-pydenticon
django-cors-headers

